[css-text] Fix tests that rely on the specific width of glyphs
The following 2 tests:
* external/wpt/css/css-text/i18n/css3-text-line-break-opclns-004.html
* external/wpt/css/css-text/i18n/css3-text-line-break-opclns-011.html

are failing on specific configurations because they assume some
characters are installed on the test machine, and they have specific
widths, but that is not the case on our try-bots.

This patch uses a [css-orientation-test] font that provides a 1em
square glyph for every Unicode code points.

[css-orientation-test]: https://github.com/adobe-fonts/css-orientation-test

Bug: 817804
Change-Id: I17a1b7d1b34b209303f8b6684644cff3d7ce62a0